+++ examples/bootstrap4/app.py
@@ -1,0 +1,202 @@
+# -*- coding: utf-8 -*-
+from flask import Flask, render_template, request, flash, Markup
+
+from flask_wtf import FlaskForm, CSRFProtect
+from wtforms import StringField, SubmitField, BooleanField, PasswordField, IntegerField, TextField,\
+    FormField, SelectField, FieldList
+from wtforms.validators import DataRequired, Length
+from wtforms.fields import *
+
+from flask_bootstrap import Bootstrap4
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.secret_key = 'dev'
+
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
+
+# set default button sytle and size, will be overwritten by macro parameters
+app.config['BOOTSTRAP_BTN_STYLE'] = 'primary'
+app.config['BOOTSTRAP_BTN_SIZE'] = 'sm'
+# app.config['BOOTSTRAP_BOOTSWATCH_THEME'] = 'lumen'  # uncomment this line to test bootswatch theme
+
+# set default icon title of table actions
+app.config['BOOTSTRAP_TABLE_VIEW_TITLE'] = 'Read'
+app.config['BOOTSTRAP_TABLE_EDIT_TITLE'] = 'Update'
+app.config['BOOTSTRAP_TABLE_DELETE_TITLE'] = 'Remove'
+app.config['BOOTSTRAP_TABLE_NEW_TITLE'] = 'Create'
+
+bootstrap = Bootstrap4(app)
+db = SQLAlchemy(app)
+csrf = CSRFProtect(app)
+
+
+class ExampleForm(FlaskForm):
+    """An example form that contains all the supported bootstrap style form fields."""
+    date = DateField(description="We'll never share your email with anyone else.")  # add help text with `description`
+    datetime = DateTimeField(render_kw={'placeholder': 'this is placeholder'})  # add HTML attribute with `render_kw`
+    image = FileField(render_kw={'class': 'my-class'})  # add your class
+    option = RadioField(choices=[('dog', 'Dog'), ('cat', 'Cat'), ('bird', 'Bird'), ('alien', 'Alien')])
+    select = SelectField(choices=[('dog', 'Dog'), ('cat', 'Cat'), ('bird', 'Bird'), ('alien', 'Alien')])
+    selectmulti = SelectMultipleField(choices=[('dog', 'Dog'), ('cat', 'Cat'), ('bird', 'Bird'), ('alien', 'Alien')])
+    bio = TextAreaField()
+    title = StringField()
+    secret = PasswordField()
+    remember = BooleanField('Remember me')
+    submit = SubmitField()
+
+
+class HelloForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired(), Length(1, 20)])
+    password = PasswordField('Password', validators=[DataRequired(), Length(8, 150)])
+    remember = BooleanField('Remember me')
+    submit = SubmitField()
+
+
+class ButtonForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired(), Length(1, 20)])
+    submit = SubmitField()
+    delete = SubmitField()
+    cancel = SubmitField()
+
+
+class TelephoneForm(FlaskForm):
+    country_code = IntegerField('Country Code')
+    area_code = IntegerField('Area Code/Exchange')
+    number = TextField('Number')
+
+
+class IMForm(FlaskForm):
+    protocol = SelectField(choices=[('aim', 'AIM'), ('msn', 'MSN')])
+    username = TextField()
+
+
+class ContactForm(FlaskForm):
+    first_name = TextField()
+    last_name = TextField()
+    mobile_phone = FormField(TelephoneForm)
+    office_phone = FormField(TelephoneForm)
+    emails = FieldList(TextField("Email"), min_entries=3)
+    im_accounts = FieldList(FormField(IMForm), min_entries=2)
+
+
+class Message(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    text = db.Column(db.Text, nullable=False)
+    author = db.Column(db.String(100), nullable=False)
+    category = db.Column(db.String(100), nullable=False)
+    draft = db.Column(db.Boolean, default=False, nullable=False)
+    create_time = db.Column(db.Integer, nullable=False, unique=True)
+
+
+@app.before_first_request
+def before_first_request_func():
+    db.drop_all()
+    db.create_all()
+    for i in range(20):
+        m = Message(
+            text=f'Test message {i+1}',
+            author=f'Author {i+1}',
+            category=f'Category {i+1}',
+            create_time=4321*(i+1)
+            )
+        if i % 4:
+            m.draft = True
+        db.session.add(m)
+    db.session.commit()
+
+
+@app.route('/')
+def index():
+    return render_template('index.html')
+
+
+@app.route('/form', methods=['GET', 'POST'])
+def test_form():
+    form = HelloForm()
+    return render_template('form.html', form=form, telephone_form=TelephoneForm(), contact_form=ContactForm(), im_form=IMForm(), button_form=ButtonForm(), example_form=ExampleForm())
+
+
+@app.route('/nav', methods=['GET', 'POST'])
+def test_nav():
+    return render_template('nav.html')
+
+
+@app.route('/pagination', methods=['GET', 'POST'])
+def test_pagination():
+    page = request.args.get('page', 1, type=int)
+    pagination = Message.query.paginate(page, per_page=10)
+    messages = pagination.items
+    return render_template('pagination.html', pagination=pagination, messages=messages)
+
+
+@app.route('/flash', methods=['GET', 'POST'])
+def test_flash():
+    flash('A simple default alert—check it out!')
+    flash('A simple primary alert—check it out!', 'primary')
+    flash('A simple secondary alert—check it out!', 'secondary')
+    flash('A simple success alert—check it out!', 'success')
+    flash('A simple danger alert—check it out!', 'danger')
+    flash('A simple warning alert—check it out!', 'warning')
+    flash('A simple info alert—check it out!', 'info')
+    flash('A simple light alert—check it out!', 'light')
+    flash('A simple dark alert—check it out!', 'dark')
+    flash(Markup('A simple success alert with <a href="#" class="alert-link">an example link</a>. Give it a click if you like.'), 'success')
+    return render_template('flash.html')
+
+
+@app.route('/table')
+def test_table():
+    page = request.args.get('page', 1, type=int)
+    pagination = Message.query.paginate(page, per_page=10)
+    messages = pagination.items
+    titles = [('id', '#'), ('text', 'Message'), ('author', 'Author'), ('category', 'Category'), ('draft', 'Draft'), ('create_time', 'Create Time')]
+    return render_template('table.html', messages=messages, titles=titles, Message=Message)
+
+
+@app.route('/table/<int:message_id>/view')
+def view_message(message_id):
+    message = Message.query.get(message_id)
+    if message:
+        return f'Viewing {message_id} with text "{message.text}". Return to <a href="/table">table</a>.'
+    return f'Could not view message {message_id} as it does not exist. Return to <a href="/table">table</a>.'
+
+
+@app.route('/table/<int:message_id>/edit')
+def edit_message(message_id):
+    message = Message.query.get(message_id)
+    if message:
+        message.draft = not message.draft
+        db.session.commit()
+        return f'Message {message_id} has been editted by toggling draft status. Return to <a href="/table">table</a>.'
+    return f'Message {message_id} did not exist and could therefore not be edited. Return to <a href="/table">table</a>.'
+
+
+@app.route('/table/<int:message_id>/delete', methods=['POST'])
+def delete_message(message_id):
+    message = Message.query.get(message_id)
+    if message:
+        db.session.delete(message)
+        db.session.commit()
+        return f'Message {message_id} has been deleted. Return to <a href="/table">table</a>.'
+    return f'Message {message_id} did not exist and could therefore not be deleted. Return to <a href="/table">table</a>.'
+
+
+@app.route('/table/<int:message_id>/like')
+def like_message(message_id):
+    return f'Liked the message {message_id}. Return to <a href="/table">table</a>.'
+
+
+@app.route('/table/new-message')
+def new_message():
+    return 'Here is the new message page. Return to <a href="/table">table</a>.'
+
+
+@app.route('/icon')
+def test_icon():
+    return render_template('icon.html')
+
+
+if __name__ == '__main__':
+    app.run(debug=True)