+++ tests/test_boostrap5/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from flask_bootstrap import Bootstrap5
+
+
+@pytest.fixture(autouse=True)
+def bootstrap(app):
+    yield Bootstrap5(app)