+++ tests/test_bootstrap4/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from flask_bootstrap import Bootstrap4
+
+
+@pytest.fixture(autouse=True)
+def bootstrap(app):
+    yield Bootstrap4(app)